+++ lib/src/client/user_context.dart
@@ -1,0 +1,11 @@
+// Copyright 2022 Kato Shinya. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+enum UserContext {
+  /// OAuth 2.0 User Context
+  oauth2,
+
+  /// OAuth 2.0 Application-Only
+  oauth2AppOnly,
+}