+++ lib/src/twitter_api.dart
@@ -3,25 +3,19 @@
 // modification, are permitted provided the conditions.
 
 // Project imports:
+import 'package:twitter_api_v2/src/client/client_context.dart';
+import 'package:twitter_api_v2/src/client/oauth_tokens.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweets_service.dart';
 import 'package:twitter_api_v2/src/service/twitter_service.dart';
 import 'package:twitter_api_v2/src/service/users/users_service.dart';
-import 'package:twitter_api_v2/src/twitter_client.dart';
 
 abstract class TwitterApi {
   /// Returns the new instance of [TwitterApi].
   factory TwitterApi({
-    required String consumerKey,
-    required String consumerSecret,
-    required String accessToken,
-    required String accessTokenSecret,
+    required String bearerToken,
+    OAuthTokens? oauthTokens,
   }) =>
-      _TwitterApi(
-        consumerKey: consumerKey,
-        consumerSecret: consumerSecret,
-        accessToken: accessToken,
-        accessTokenSecret: accessTokenSecret,
-      );
+      _TwitterApi(bearerToken: bearerToken, oauthTokens: oauthTokens);
 
   /// Returns the tweets service.
   TweetsService get tweetService;
@@ -32,16 +26,12 @@ abstract class TwitterApi {
 
 class _TwitterApi implements TwitterApi {
   _TwitterApi({
-    required String consumerKey,
-    required String consumerSecret,
-    required String accessToken,
-    required String accessTokenSecret,
+    required String bearerToken,
+    OAuthTokens? oauthTokens,
   }) : _twitterService = TwitterService(
-          client: TwitterClient(
-            consumerKey: consumerKey,
-            consumerSecret: consumerSecret,
-            accessToken: accessToken,
-            accessTokenSecret: accessTokenSecret,
+          context: ClientContext(
+            bearerToken: bearerToken,
+            oauthTokens: oauthTokens,
           ),
         );